+++ metric_consolidator.go
@@ -82,11 +82,13 @@ func (mc *MetricConsolidator) Flush(ctx context.Context) {
 	}
 }
 
-// ReceiveMetric will push a Metric in to one of the MetricMaps
-func (mc *MetricConsolidator) ReceiveMetric(m *Metric) {
-	m.TagsKey = m.FormatTagsKey()
+// ReceiveMetrics will push a slice of Metrics in to one of the MetricMaps
+func (mc *MetricConsolidator) ReceiveMetrics(metrics []*Metric) {
 	mmTo := <-mc.maps
-	mmTo.Receive(m)
+	for _, m := range metrics {
+		m.TagsKey = m.FormatTagsKey()
+		mmTo.Receive(m)
+	}
 	mc.maps <- mmTo
 }